requirements.txt
beautifulsoup4==4.10.0
matplotlib==3.5.1
newsapi==0.1.1
newsapi_python==0.2.6
newsdataapi==0.1.1
nltk==3.6.5
numpy==1.21.4
pandas==1.3.4
pymongo==3.12.1
python_dateutil==2.8.2
requests==2.26.0
selenium==4.1.0
TA_Lib==0.4.24
webdriver_manager==3.5.2
yfinance==0.1.67
